Update styling for "browse my ballot" page
This page allows the user to enter an address for candidates they would
like to browse. After entering a search. they will be sent to a new
page to see their results with further details. Which inputs are presented
to the user is controlled by the selection of one of the area buttons,
as decribed in issue #43

Also in this commit:

 - resized the navi_* icons in Inkscape.
 - some basic form validation is included (all the fields that are shown
   are considered required).