Style the browse my ballot results page
Added a new route and controller for the browse results. It uses the same
basic styling for the race selection as was used on the browse search page.
Not sure why the font size isn't getting inherited to all the children, so
it's duplicated in a number of spots.

The selected versions of the city, national, and state icons were missing,
so I edited the existing gray ones using the same color scheme as the
existing selected icons (per suggestion from ninjascience). This implements
most of issue #53.