safari fix when including SVG on front page
Apparently, the SVG standard actually says that when an SVG element doesn't
specify a size, it should take up the entire available space in its parent.
Safari seems to interpret that by implicitly setting the height to 100%,
(the width is explicitly constrained by the bootstrap properties). This
takes up too much vertical space.

I'm trying to just explicitly set a parent div to 200px. thinking that
will be larger than the width of a (sub) col-md-4 for even very large screens
and won't affect the other browsers (tested firefox and chrome) too much.