finish a cleaned-up/working spending visualization
showing categories and relative amouts as both a bar chart or a pie chart.
The pie chart may need work to tease apart the labels when several very
small slices end up next to each other. For both charts, I went with a
tooltip that shows up onmouseover because we don't have images for all the
possible categories.