clean up the interaction of the donorList
I determined that my understanding of overflow and height was wrong, that
an explicit height with overflow: initial does NOT mean it should push
later elements down. (dispite safari doing so anyway). I still set the
height for the div so it can clip the rest of the list, but when we show
the full list I also set the height back to auto.

As part of the cleanup, I created a 'wider' class that removes the padding
that Bootstrap adds each time I use a 'col-md-*' class, which happens a
lot with the donorList. The resultant class reads like this

  class="wider col-md-12"

Finally, the div that switches between clipping the list and revealing it
no longer includes the header. The header serves as an alternative button
to toggle showing the list and is a different height than the list entries.
Without the header, it is much easier to compute a value for the div
height that doesn't cut off an entry.