change the orientation of the "Money Spent" bar chart to show category names
Based on a suggestion from John, this changes the bars in the bar chart to
be horizontal so that adding more categories doesn't cause the space
available for the labels to get crowded. With more categories, the chart
simply grows down rather than having the bars get narrower. Because we
are showing all the reported categories and the dollar amounts, there is
no longer a need for the tooltip (although it may be useful elsewhere)