Cobble together a mostly working "View all Oregon" page
All the endpoints are different, so there were only limited places we
could share code with the campaign page. I still started by trying to
wegde the code into the CampaignService anyway (not realizing how much
the endpoints were different), and munge the data once we receive it
into the format expected by the various directives.

Among the differences: instead of having the backend send all the
transactions and tally them in the browser, there are now two new endpoints
which summarize all the contributions and expenditures for all of Oregon.
(It also seems that the category strings returned were still slightly
different between the summary and the individual candidates?)

Another difference is in how the spent dollars are computed when multiple
purpose codes are indicated. Rather than double- (or triple-) count the
dollars by including the full amount in all given categories, we spread the
total amount evenly across all given purpose_codes. This means that we
should expect the sum of all the sub-categories to match the total dollars
spent. This matches how the backend will compute the summaries for the
Oregon Summary page.

Although Sam assures me it will be a simple computation, we don't yet have
the "Top Contributors" for either individuals (or businesses or PACs) and
we can't compute them (since we don't have access to all the transactions).
Consequently, I hide the top donor lists when they are empty.

Finally, due to the much larger dollar amounts on the Oregon page, I added
another step to the state-by-state directive's output range.