Extend donor lists to include PACs, and link to PAC pages if available
When committees show up in a donor list and we have a detail page for them,
the name is now a link to that donor page!

The layout really only accomodates two donor lists, so I have to hide one
of them. I resurrected the arrow images from an early set of site designs
and put them on either side of the lists, and caused them to cycle the
donor lists in different directions. This meant I needed to put the donor
lists into a list of their own so the controller could shuffle them, as
well as hide the third one easily. (At the same time being sure to not
place an empty list in there with just a name. No one wants to see an empty
list.) I hide the arrows on narrow screens, since all three lists are
shown in that case. However, I also discovered that ngHide does not
override the bootstrap !important display class for .xs-hidden, so I had
to use ngIf instead (... so many hacks, ugh)

I felt that the  donor list item's donation bar ought to be on a scale
from zero to the max donation, too. My reasoning is that if there are only
two entries (say $2000 and $1000), the smaller one should be roughly half
the size of the larger one. If the scale starts at the smallest donation
instead of zero, nothing shows up for the smaller donation, regardless of
how close they may be.