plumb the data for top PACs to the "View all Oregon" page
All the pieces were there, just needed to be hooked up.
One small comment: I resolved the 3 donor lists together in a $q.all
because simply setting the fields in the donor object didn't seem to
trigger the $watch on it. Creating a new object with all new fields
did the trick.